BOOKING-134: Refactor after latest develop changes merged
