+++ config/api_platform/resource.yaml
@@ -1,3 +1,3 @@
 ---
 resources:
-  App\Entity\Resource:
+  App\Entity\Main\Resource:
