+++ config/api_platform/user-booking.yaml
@@ -1,6 +1,6 @@
 ---
 resources:
-  App\Entity\UserBooking:
+  App\Entity\Main\UserBooking:
     itemOperations:
       get: ~
     collectionOperations:
