+++ src/Entity/Main/BookingDetails.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity;
+namespace App\Entity\Main;
 
 use ApiPlatform\Core\Annotation\ApiProperty;
