+++ tests/Api/BookingTest.php
@@ -3,7 +3,7 @@
 namespace App\Tests\Api;
 
 use App\Entity\ApiKeyUser;
-use App\Entity\Booking;
+use App\Entity\Main\Booking;
 use App\Message\CreateBookingMessage;
 use App\Message\WebformSubmitMessage;
 use App\MessageHandler\CreateBookingHandler;
